6/3/25
Added the door animation, and finished the second video. I think instead of an fps game, I am going to create a parkour game, with like ten levels, and then it keeps track of the time from when you started. Also it keeps track of the all time best score